% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFcor_plot.R
\name{CDFcor_plot}
\alias{CDFcor_plot}
\title{CDFcor_plot}
\usage{
CDFcor_plot(x, y, grid, xlim, ylim)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{grid}{regulates the grid size (which is grid^2) of the equal-spaced grid}

\item{xlim}{limits for the x-axis, defaults to (min(x),max(x))}

\item{ylim}{limits for the y-axis, defaults to (min(y),max(y))}
}
\value{
plot of the CDF correlation of x and y
}
\description{
plots the CDF correlation of two vectors x and y on an equal-spaced grid
}
