% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_bound_quantile.R
\name{FH_bound_quantile}
\alias{FH_bound_quantile}
\title{FH_bound_quantile}
\usage{
FH_bound_quantile(x, y, alpha, beta, upper = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{alpha}{quantile level for x}

\item{beta}{quantile level for y}

\item{upper}{if TRUE (default) the upper bound is computed, if FALSE the lower bound}
}
\value{
Fréchet-Hoeffding upper/lower bound of x and y at alpha and beta
}
\description{
computes the Fréchet-Hoeffding upper/lower bound of x and y at quantile levels alpha and beta
}
