% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_bound_threshold.R
\name{FH_bound_threshold}
\alias{FH_bound_threshold}
\title{FH_bound_threshold}
\usage{
FH_bound_threshold(x, y, a, b, upper = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{a}{threshold for x}

\item{b}{threshold for y}

\item{upper}{if TRUE (default) the upper bound is computed, if FALSE the lower bound}
}
\value{
Fréchet-Hoeffding upper/lower bound of x and y at a and b
}
\description{
computes the Fréchet-Hoeffding upper/lower bound of x and y at thresholds a and b
}
