% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdep.R
\name{mdep-package}
\alias{mdep-package}
\alias{mdep}
\alias{gcor}
\alias{gdis}
\alias{pscore}
\title{Estimate mutual dependency and related measures}
\usage{
gcor(x, y = NULL, k = NULL, data = NULL, drop = TRUE, use = "everything")

gdis(x, k = NULL, use = "everything", ...)

pscore(x, y = NULL, k = NULL, data = NULL, drop = TRUE, use = "everything")
}
\arguments{
\item{x}{a vector, matrix, data frame or formula. If formula, \code{data} should be specified.
\code{gdis} requires a matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with compatible dimensions to \code{x}.}

\item{k}{\code{NULL} (default) or an integer specifying the number of groups for discretization.
Numerical data are divided into \code{k} groups using \code{k}-quantiles.
If \code{NULL}, it is determined automatically.}

\item{data}{\code{NULL} (default) or a data frame. Required if \code{x} is a formula.}

\item{drop}{a logical. If \code{TRUE}, the returned value is coerced to
a vector when one of its dimensions is one.}

\item{use}{a character specifying how to handle missing values.
It should be (an abbreviation of) one of "everything" (default.
Treats missing values as a single observed categorical value, namely \code{NA}),
"complete.obs" (casewise deletion; rows containing any missing value are removed),
or "pairwise.complete.obs" (pairwise deletion; for any pair (x,y), the i-th row
is removed if x[i] or y[i] is missing. It does not affect another pair, e.g. (w,z)).}

\item{...}{additional arguments (\code{diag} and \code{upper}) passed to \code{as.dist} function.
See \link{as.dist} for details.}
}
\value{
For \code{gcor} and \code{pscore}, a numeric matrix is returned (or a vector if \code{drop = TRUE}).
For \code{gdis}, an object of class \code{"dist"} is returned.
}
\description{
Estimate generalized correlation measure (\code{gcor}),
generalized dissimilarity between variables (\code{gdis}), and
predictability score (\code{pscore}) based on mutual dependency.
}
\examples{
# Generalized correlation measure
gcor(iris)

# Clustering
gd <- gdis(iris)
hc <- hclust(gd, method = "ward.D2")
plot(hc)

# Multidimensional scaling
mds <- cmdscale(gd, k = 2)
plot(mds, type = "n", xlab = "", ylab = "", asp = 1, axes = FALSE,
     main = "cmdscale with gdis(iris)")
text(mds[,1], mds[,2], rownames(mds))

# Predictability of Species from other variables
ps <- pscore(Species ~ ., data = iris)
dotchart(sort(ps), xlim = c(0, 1), main = "Predictability of Species")
}
\references{
Suzuki, R. (2025). \emph{A generalization of correlation coefficient}. preprint.
\url{https://r-suzuki.github.io/preprints}
}
