% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cor.R
\name{mean_cor}
\alias{mean_cor}
\title{mean_cor}
\usage{
mean_cor(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}
}
\value{
mean correlation of x and y
}
\description{
computes the mean correlation of two vectors x and y
}
