% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{plot_contour}
\alias{plot_contour}
\title{contour_plot}
\usage{
plot_contour(f, xlim, ylim, grid, xlab, ylab, bins)
}
\arguments{
\item{f}{a function in two variables, x and y}

\item{xlim}{limits for the x-axis}

\item{ylim}{limits for the y-axis}

\item{grid}{regulates the grid size (which is grid^2) of the equal-spaced grid}

\item{bins}{number of colour bins}

\item{xlab:}{label for the x-axis}

\item{ylab:}{label for the y-axis}
}
\value{
contour plot of f
}
\description{
visualizes a two-dimensional function with a contour plot on an equal-spaced grid with a binned colour scheme
}
