% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qcor.R
\name{Qcor}
\alias{Qcor}
\title{Qcor}
\usage{
Qcor(x, y, alpha, beta)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{alpha}{quantile level for x}

\item{beta}{quantile level for y}
}
\value{
quantile correlation of x and y at alpha and beta
}
\description{
computes the quantile correlation of two vectors x and y at quantile levels alpha and beta
}
