% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qcov.R
\name{Qcov}
\alias{Qcov}
\title{Qcov}
\usage{
Qcov(x, y, alpha, beta)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{alpha}{quantile level for x}

\item{beta}{quantile level for y}
}
\value{
quantile covariance of x and y at alpha and beta
}
\description{
computes the quantile covariance of two vectors x and y at quantile levels alpha and beta
}
