% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot}
\usage{
scatter_plot(
  x,
  y,
  bubbleplot = FALSE,
  xlab,
  ylab,
  xlim = c(min(x), max(x)),
  ylim = c(min(y), max(y))
)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{bubbleplot}{if TRUE uses a bubbleplot, which is useful to avoid overplotting for discrete variables, if FALSE (the default) draws a usual scatter plot}

\item{xlab:}{label for the x-axis}

\item{ylab:}{label for the y-axis}

\item{xlim:}{limits for the x-axis, defaults to (min(x),max(x))}

\item{ylim:}{limits for the y-axis, defaults to (min(y),max(y))}
}
\value{
scatter plot of x and y
}
\description{
draws a scatter plot (or a bubble plot) of x and y
}
