% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tcor.R
\name{Tcor}
\alias{Tcor}
\title{Tcor}
\usage{
Tcor(x, y, a, b)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of the same length as x}

\item{a}{threshold for x}

\item{b}{threshold for y}
}
\value{
threshold correlation of x and y at a and b
}
\description{
computes the threshold correlation of two vectors x and y at thresholds a and b
}
