% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdep.R
\name{gcor}
\alias{gcor}
\alias{gdis}
\alias{pscore}
\alias{mdep}
\title{Estimate mutual dependencies and related measures}
\usage{
gcor(x, y = NULL, k = NULL, data = NULL, drop = TRUE)

gdis(x, y = NULL, k = NULL, data = NULL)

pscore(x, y = NULL, k = NULL, data = NULL, drop = TRUE)

mdep(x, y = NULL, k = NULL, measure, data = NULL, drop = FALSE)
}
\arguments{
\item{x}{a vector, matrix, data frame or formula. If formula, \code{data} should be specified.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with compatible dimensions to \code{x}.}

\item{k}{\code{NULL} (default) or an integer specifying the number of groups for discretization.
Numerical data are divided into \code{k} groups using \code{k}-quantiles.
If \code{NULL}, it is determined automatically.}

\item{data}{\code{NULL} (default) or a data frame. Required if \code{x} is a formula.}

\item{drop}{a logical. If \code{TRUE}, the returned value is coerced to
a vector when one of its dimensions is one.}

\item{measure}{a character specifying the type of measure, one of \code{"cor"}, \code{"dist"}, \code{"pred"}.
\code{gcor} is a wrapper for \code{mdep} with \code{measure = "cor"}.
Similarly, \code{gdis} wraps \code{measure = "dist"}, and \code{pscore} wraps \code{measure = "pred"}.}
}
\value{
For \code{gcor} and \code{pscore}, a numeric matrix is returned (or a vector if \code{drop = TRUE}).
For \code{gdis}, an object of class \code{"dist"} is returned.
}
\description{
Estimate generalized correlation measure, generalized distance between variables,
and predictability score based on mutual dependencies.
}
\examples{
# Generalized correlation measure
gcor(iris)

# Clustering
gd <- gdis(iris)
hc <- hclust(gd, method = "ward.D2")
plot(hc)

# Multidimensional scaling
mds <- cmdscale(gd, k = 2)
plot(mds, type = "n", xlab = "", ylab = "", asp = 1, axes = FALSE,
     main = "cmdscale with gdis(iris)")
text(mds[,1], mds[,2], rownames(mds))

# Predictability of Species from other variables
ps <- pscore(Species ~ ., data = iris)
dotchart(sort(ps), xlim = c(0, 1), main = "Predictability of Species")
}
\references{
Suzuki, R. (2025). \emph{A generalization of correlation coefficient}. preprint.
\url{https://r-suzuki.github.io/preprints}
}
