% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_contour}
\alias{plot_contour}
\title{plot_contour}
\usage{
plot_contour(
  f,
  xlim = c(min(x), max(x)),
  ylim = c(min(y), max(y)),
  grid,
  xlab,
  ylab,
  bins
)
}
\arguments{
\item{f}{a function in two variables, x and y}

\item{xlim}{limits for the x-axis, defaults to (min(x),max(x))}

\item{ylim}{limits for the y-axis, defaults to (min(y),max(y))}

\item{grid}{regulates the grid size (which is grid^2) of the equal-spaced grid}

\item{bins}{number of colour bins}

\item{xlab:}{label for the x-axis}

\item{ylab:}{label for the y-axis}
}
\value{
contour plot of f
}
\description{
visualizes a two-dimensional function with a contour plot on an equal-spaced grid with a binned colour scheme
}
